подправила объединение модулей
