+++ src/components/film.js
@@ -1,20 +1,11 @@
 export const createFilmTemplate = () => {
   return (
-    `<section class="films"></section>`
-  );
-};
-
-export const createFilmListTemplate = () => {
-  return (
-    `<section class="films-list">
-      <h2 class="films-list__title visually-hidden">All movies. Upcoming</h2>
-      <div class="films-list__container"></div>
+    `<section class="films"></section>
+      <section class="films-list">
+        <h2 class="films-list__title visually-hidden">All movies. Upcoming</h2>
+        <div class="films-list__container"></div>
+      </section>
+      <button class="films-list__show-more">Show more</button>
     </section>`
-  );
-};
-
-export const createShowButtonTemplate = () => {
-  return (
-    `<button class="films-list__show-more">Show more</button>`
   );
 };
