+++ src/main.js
@@ -1,22 +1,20 @@
 
+import {createProfileTemplate} from "./components/profile.js";
+import {createMenuTemplate} from "./components/menu.js";
+import {createSortingTemplate} from "./components/sorting.js";
+import {createFilmTemplate} from "./components/film.js";
+import {createFilmCardTemplate} from "./components/filmCard.js";
+import {createTopRatedTemplate} from "./components/topRated.js";
+import {createMostCommentedTemplate} from "./components/mostCommented.js";
+import {createStatisticsTemplate} from "./components/statistics.js";
+import {createPopupTemplate} from "./components/popup.js";
+
 const FILM_CARDS = 5;
 const EXTRA_FILM_CARDS = 2;
 
 const header = document.querySelector(`.header`);
 const main = document.querySelector(`.main`);
 const body = document.querySelector(`body`);
-
-import {createProfileTemplate} from "./components/profile.js";
-import {createMenuTemplate} from "./components/menu.js";
-import {createSortingTemplate} from "./components/menu.js";
-import {createFilmTemplate} from "./components/film.js";
-import {createFilmListTemplate} from "./components/film.js";
-import {createFilmCardTemplate} from "./components/filmCard.js/index.js";
-import {createTopRatedTemplate} from "./components/topRated.js/index.js";
-import {createMostCommentedTemplate} from "./components/mostCommented.js/index.js";
-import {createShowButtonTemplate} from "./components/film.js";
-import {createStatisticsTemplate} from "./components/statistics.js";
-import {createPopupTemplate} from "./components/popup.js";
 
 const render = (container, template, place) => {
   container.insertAdjacentHTML(place, template);
@@ -31,7 +29,6 @@ render(main, createFilmTemplate(), `beforeend`);
 // Отрисовка главных блоков
 const films = document.querySelector(`.films`);
 
-render(films, createFilmListTemplate(), `beforeend`);
 render(films, createTopRatedTemplate(), `beforeend`);
 render(films, createMostCommentedTemplate(), `beforeend`);
 
@@ -43,10 +40,6 @@ const filmsListContainer = filmsList.querySelector(`.films-list__container`);
 for (let i = 0; i < FILM_CARDS; i++) {
   render(filmsListContainer, createFilmCardTemplate(), `beforeend`);
 }
-
-// Отрисовка кнопки "Показать еще"
-render(filmsList, createShowButtonTemplate(), `beforeend`);
-
 
 // Отрисовка доп карточек (topRated и mostCommented)
 const filmListExtraContainer = main.querySelectorAll(`.films-list--extra`);
